Change: Refactor the pontos-changelog CLI
Only print the generated changelog to stdout and make writing into a
file optional. Behave similar as the prepare command where the current
release version is gathered from the last git tag if it isn't passed as
argument. This allows to generate the same changelog output as prepare
would create.
